changes file organized
